Year_1/1_Winter_Term/Introduction_to_Computer_Security_-_I-UPB/Assignments/I-UPB_Assignment_2/requirements.txt
pycryptodome==3.11.0
pyminizip==0.2.4
termcolor==1.1.0
